+++ src/features/game/components/common/index.ts
@@ -1,0 +1,1 @@
+export { MatchBoard } from './MatchBoard';
