--- src/features/game/components/index.ts
@@ -1,1 +1,0 @@
-export { CarouselCard } from './carousel/CarouselCard';
