+++ src/features/game/components/watch-point/index.ts
@@ -1,3 +1,2 @@
-export { MatchBoard } from './MatchBoard';
 export { StartingPitcherTable } from './StartingPitcherTable';
 export { TeamLineup } from './TeamLineup';
